Made RoutingResults public
Server.internalPublish returns the RoutingResults, but
PostOffice.RoutingResults was not public. Since PostOffice itself is not
public either, RoutingResults had to be moved out of this class.
