+++ broker/src/main/java/io/moquette/broker/Server.java
@@ -306,7 +306,7 @@ public class Server {
      * @throws IllegalStateException if the integration is not yet started
      * @return
      */
-    public PostOffice.RoutingResults internalPublish(MqttPublishMessage msg, final String clientId) {
+    public RoutingResults internalPublish(MqttPublishMessage msg, final String clientId) {
         final int messageID = msg.variableHeader().packetId();
         if (!initialized) {
             LOG.error("Moquette is not started, internal message cannot be published. CId: {}, messageId: {}", clientId,
@@ -314,7 +314,7 @@ public class Server {
             throw new IllegalStateException("Can't publish on a integration is not yet started");
         }
         LOG.trace("Internal publishing message CId: {}, messageId: {}", clientId, messageID);
-        final PostOffice.RoutingResults routingResults = dispatcher.internalPublish(msg);
+        final RoutingResults routingResults = dispatcher.internalPublish(msg);
         msg.payload().release();
         return routingResults;
     }
